chron: About the Friend 친구 삭제 - 친구 상태 검증 추가

+++ src/main/java/com/api/stuv/domain/friend/service/FriendService.java
@@ -57,6 +57,7 @@ public class FriendService {
     @Transactional
     public void deleteFriend(Long userId, Long friendId) {
         Friend friend = friendRepository.findById(friendId).orElseThrow(() -> new NotFoundException(ErrorCode.FRIEND_NOT_FOUND));
+        if ( !friend.getStatus().equals(FriendStatus.ACCEPTED) ) throw new NotFoundException(ErrorCode.FRIEND_NOT_FOUND);
         if ( !(friend.getUserId().equals(userId) || friend.getFriendId().equals(userId)) ) throw new AccessDeniedException(ErrorCode.FRIEND_DELETE_NOT_AUTHORIZED);
         friendRepository.delete(friend);
     }
